refactor: reorganize cache implementation
